requirements.txt
flask
pillow
numpy
scikit-image
tensorflow
flask-cors
gunicorn
tensorflow-cpu